+++ src/AsyncHTTPRequest_Impl_Generic.h
@@ -18,7 +18,7 @@
   You should have received a copy of the GNU General Public License along with this program.
   If not, see <https://www.gnu.org/licenses/>.
 
-  Version: 1.12.0
+  Version: 1.13.0
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -37,6 +37,7 @@
   1.10.2   K Hoang     09/11/2022 Default to reconnect to the same host:port after connected for new HTTP sites
   1.11.0   K Hoang     28/11/2022 Add support to ESP32 boards using LwIP ENC28J60 Ethernet
   1.12.0   K Hoang     30/11/2022 Add support to ESP32 boards using LwIP W5500 Ethernet. Fix bug
+  1.13.0   K Hoang     01/02/2023 Add support to ESP32 using LwIP W6100 Ethernet. Fix _parseURL() bug
  *****************************************************************************************************************************/
 
 #pragma once
@@ -1198,12 +1199,29 @@ bool  AsyncHTTPRequest::_parseURL(const String& url)
   }
 
   int pathBeg = url.indexOf('/', hostBeg);
-
+  
+  int hostEnd;
+  int portBeg;
+  
   if (pathBeg < 0)
-    return false;
+  {
+    if ( url.indexOf(':', hostBeg) < 0 )
+    {
+      // No port, just https://www.aaa.com
+      hostEnd = url.length();
+    }
+    else
+    {
+      // with port, https://www.aaa.com:443
+      hostEnd = url.indexOf(':', hostBeg);
+    }
+  }
+  else
+  {
+    hostEnd = pathBeg;
+  }
 
-  int hostEnd = pathBeg;
-  int portBeg = url.indexOf(':', hostBeg);
+  portBeg = url.indexOf(':', hostBeg);
 
   if (portBeg > 0 && portBeg < pathBeg)
   {