+++ changelog.md
@@ -10,12 +10,14 @@
 <a href="https://www.buymeacoffee.com/khoihprog6" title="Donate to my libraries using BuyMeACoffee"><img src="https://cdn.buymeacoffee.com/buttons/v2/default-yellow.png" alt="Donate to my libraries using BuyMeACoffee" style="height: 50px !important;width: 181px !important;" ></a>
 <a href="https://www.buymeacoffee.com/khoihprog6" title="Donate to my libraries using BuyMeACoffee"><img src="https://img.shields.io/badge/buy%20me%20a%20coffee-donate-orange.svg?logo=buy-me-a-coffee&logoColor=FFDD00" style="height: 20px !important;width: 200px !important;" ></a>
 
+
 ---
 ---
 
 ## Table of Contents
 
 * [Changelog](#changelog)
+	* [Releases v1.13.0](#releases-v1130)
 	* [Releases v1.12.0](#releases-v1120)
 	* [Releases v1.11.0](#releases-v1110)
 	* [Releases v1.10.2](#releases-v1102)
@@ -50,6 +52,12 @@
 ---
 
 ## Changelog
+
+### Releases v1.13.0
+
+1. Add support to ESP32 boards using `LwIP W6100 Ethernet`
+2. Fix bug of `_parseURL()`. Check [Bug with _parseURL() #21](https://github.com/khoih-prog/AsyncHTTPSRequest_Generic/issues/21)
+3. Improve `README.md` so that links can be used in other sites, such as `PIO`
 
 ### Releases v1.12.0